requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
stripe==7.8.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
python-decouple==3.8
